first pass at refactoring base class out of GCSDirectory
